+++ packages/design-system/src/hooks/index.ts
@@ -1,20 +1,36 @@
-export { useCollectionRender } from './use-collection-render';
-export { useContextProps } from './use-context-props';
+/*
+ * Copyright 2025 Hypergiant Galactic Systems Inc. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+export { useCollectionRender } from './use-collection-render/use-collection-render';
+export { useContextProps } from './use-context-props/use-context-props';
 export {
   DefaultsProvider,
   useDefaultProps,
   useDefaults,
-  type DefaultsContext,
-  type DefaultsProviderProps,
-} from './use-defaults';
-export { usePropagatingPress } from './use-propagating-press';
-export { useSlot } from './use-slot';
+} from './use-defaults/use-defaults';
+export type {
+  DefaultsContext,
+  DefaultsProviderProps,
+} from './use-defaults/types';
+export { usePropagatingPress } from './use-propagating-press/use-propagating-press';
+export { useSlot } from './use-slot/use-slot';
 export {
   ThemeProvider,
   useTheme,
-  type ThemeContext,
-  type ThemeProviderProps,
-  type ThemeVars,
-} from './use-theme';
-export { useTree } from './use-tree';
-export { useUpdateEffect } from './use-update-effect';
+} from './use-theme/use-theme';
+export type {
+  ThemeContext,
+  ThemeProviderProps,
+  ThemeVars,
+} from './use-theme/types';
+export { useTree } from './use-tree/use-tree';
+export { useUpdateEffect } from './use-update-effect/use-update-effect';
