--- packages/design-system/src/hooks/use-collection-render/index.ts
@@ -1,2 +1,0 @@
-// __private-exports
-export { useCollectionRender } from './use-collection-render';
