--- packages/design-system/src/hooks/use-context-props/index.ts
@@ -1,2 +1,0 @@
-// __private-exports
-export { useContextProps } from './use-context-props';
