+++ packages/design-system/src/hooks/use-context-props/use-context-props.ts
@@ -1,3 +1,15 @@
+/*
+ * Copyright 2025 Hypergiant Galactic Systems Inc. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
 import { mergeRefs, useObjectRef } from '@react-aria/utils';
 import {
   type Context,
@@ -10,7 +22,7 @@ import {
   type SlotProps,
   useSlottedContext,
 } from 'react-aria-components';
-import { mergeProps } from '../../utils';
+import { mergeProps } from '../../utils/props';
 
 /**
  * Reimplementation of React Aria's useContextProps, to utilize our own
