--- packages/design-system/src/hooks/use-defaults/index.ts
@@ -1,3 +1,0 @@
-// __private-exports
-export { DefaultsProvider, useDefaultProps, useDefaults } from './use-defaults';
-export { type DefaultsContext, type DefaultsProviderProps } from './types';
