--- packages/design-system/src/hooks/use-propagating-press/index.ts
@@ -1,2 +1,0 @@
-// __private-exports
-export { usePropagatingPress } from './use-propagating-press';
