--- packages/design-system/src/hooks/use-slot/index.ts
@@ -1,2 +1,0 @@
-// __private-exports
-export { useSlot } from './use-slot';
