--- packages/design-system/src/hooks/use-theme/index.ts
@@ -1,7 +1,0 @@
-// __private-exports
-export {
-  type ThemeContext,
-  type ThemeProviderProps,
-  type ThemeVars,
-} from './types';
-export { ThemeProvider, useTheme } from './use-theme';
