--- packages/design-system/src/hooks/use-tree/index.ts
@@ -1,2 +1,0 @@
-// __private-exports
-export { useTree } from './use-tree';
