--- packages/design-system/src/hooks/use-update-effect/index.ts
@@ -1,2 +1,0 @@
-// __private-exports
-export { useUpdateEffect } from './use-update-effect';
